Day 2: Use substring instead of arrays for part 2
Doing the conversion to an array to use each_slice ends up being
expensive. Manually looping over the substrings ends up being cheaper,
and not too difficult to read.

This also adds the optimizations of only checking substring lenghts up
to half the string, and that divide the string evenly.